Create DecodeDemo
this is an easy demo to show how the decode function works, the vacation id numbers directly correlate to the city names. I added a statement at the end to be a default message when there are more vacation ids than the scope of what is being called for. When the values exceed the scope, the default response will be 'Hey...This is a top 10 List :). Decoding the values make it easier to write a database with placeholders, then define the concrete values at a later time.